Grouped socket_io event handlers
